+++ source/ChanSort.Loader.Panasonic/Serializer.cs
@@ -309,6 +309,7 @@ namespace ChanSort.Loader.Panasonic
 
       this.Features.ChannelNameEdit = ChannelNameEditMode.None; // due to the chaos with binary data inside the "sname" string column, writing back a name has undesired side effects
       this.Features.CanHaveGaps = false;
+      this.Features.EncryptedFlagEdit = true;
       this.DataRoot.SortedFavorites = true;
       
       this.DataRoot.AddChannelList(this.avbtChannels);