+++ source/ChanSort.Loader.SamsungJ/DbSerializerPlugin.cs
@@ -4,9 +4,9 @@ namespace ChanSort.Loader.SamsungJ
 {
   public class DbSerializerPlugin : ISerializerPlugin
   {
-    public string PluginName { get { return "Samsung J-Series"; } }
-    public string FileFilter { get { return "channel_list_t*.zip"; } }
-    
+    public string PluginName => "Samsung J-Series (*.zip)";
+    public string FileFilter => "*.zip"; // "channel_list_t*.zip";
+
     public SerializerBase CreateSerializer(string inputFile)
     {
       return new DbSerializer(inputFile);