+++ source/ChanSort.Loader.Toshiba/DbSerializerPlugin.cs
@@ -4,9 +4,9 @@ namespace ChanSort.Loader.Toshiba
 {
   public class DbSerializerPlugin : ISerializerPlugin
   {
-    public string PluginName { get { return "Toshiba *.zip"; } }
-    public string FileFilter { get { return "*.zip"; } }
-    
+    public string PluginName => "Toshiba *.zip";
+    public string FileFilter => "Hotel*.zip";
+
     public SerializerBase CreateSerializer(string inputFile)
     {
       return new DbSerializer(inputFile);